mosdac/requirements.txt
# Core AI/ML Libraries
openai>=1.6.1
langchain==0.0.350
langchain-openai>=0.0.2
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2

# Web Scraping & Processing
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
scrapy==2.11.0
lxml==4.9.3

# Knowledge Graph & Database
neo4j==5.14.1
rdflib==7.0.0
networkx==3.2.1
py2neo==2021.2.4

# Geospatial Processing
geopandas==0.14.1
shapely==2.0.2
folium==0.15.1
pyproj==3.6.1

# NLP & Text Processing
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1
gensim==4.3.2

# Web Framework & API
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
flask==3.0.0

# Data Processing
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
loguru==0.7.2
tqdm==4.66.1

# Development & Testing
pytest==7.4.3
black==23.11.0
flake8==6.1.0
